refactor/ refactor code and rename some components and change the structure of index and _app files
